% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumProdCol}
\alias{cumProdCol}
\title{Calculate the cumulative product of each column in a matrix}
\usage{
cumProdCol(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A matrix containing the cumulative product of each column of the matrix \code{x}
}
\description{
Calculate the cumulative product of each column in a matrix
}
\examples{
a <- matrix(1:9, 3)
cumProdCol(a)
}

