% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumProdRow}
\alias{cumProdRow}
\title{Calculate the cumulative product of each row in a matrix}
\usage{
cumProdRow(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A matrix containing the cumulative product of each row of the matrix \code{x}
}
\description{
Calculate the cumulative product of each row in a matrix
}
\examples{
a <- matrix(1:9, 3)
cumProdRow(a)
}

