% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumSumRow}
\alias{cumSumRow}
\title{Calculate the cumulative sum of each row in a matrix}
\usage{
cumSumRow(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A matrix containing the cumulative sum of each row of the matrix \code{x}
}
\description{
Calculate the cumulative sum of each row in a matrix
}
\examples{
a <- matrix(1:9, 3)
cumSumRow(a)
}

