% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prodCol}
\alias{prodCol}
\title{Calculate the product of each column in a matrix}
\usage{
prodCol(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A vector containing the product of each column of the matrix \code{x}
}
\description{
Calculate the product of each column in a matrix
}
\examples{
a <- matrix(1:9, 3)
prodCol(a)
}

