% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prodRow}
\alias{prodRow}
\title{Calculate the product of each row in a matrix}
\usage{
prodRow(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A vector containing the product of each row of the matrix \code{x}
}
\description{
Calculate the product of each row in a matrix
}
\examples{
a <- matrix(1:9, 3)
prodRow(a)
}

