% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prodRowLimited}
\alias{prodRowLimited}
\title{Calculate the limited cumulative product of each row in a matrix.}
\usage{
prodRowLimited(x, limits)
}
\arguments{
\item{x}{A numeric matrix}

\item{limits}{A vector showing the number of values to be multiplied in each row, starting at the left side}
}
\value{
A matrix containing the cumulative product of each row of the matrix \code{x}, with limited values in each row used
}
\description{
Calculate the limited cumulative product of each row in a matrix.
}
\examples{
a <- matrix(1:9, 3)
prodRowLimited(a, c(1,2,3))
}

