% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{removeCols}
\alias{removeCols}
\title{Remove set a of named columns from a data frame or matrix}
\usage{
removeCols(data, cols)
}
\arguments{
\item{data}{data frame or matrix}

\item{cols}{Vector containing the names of columns}
}
\value{
A data frame or matrix with the designatd columsn removed
}
\description{
Remove set a of named columns from a data frame or matrix
}

