% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{removeRowsNA}
\alias{removeRowsNA}
\title{Remove rows from database where any row contains an NA}
\usage{
removeRowsNA(data)
}
\arguments{
\item{data}{Data frame}
}
\description{
Remove rows from database where any row contains an NA
}

