% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{repCols}
\alias{repCols}
\title{repeats each column of a matrix a specified number of times}
\usage{
repCols(x, each)
}
\arguments{
\item{x}{Matrix}

\item{each}{Scalar or a vector.  This represents the number of times each column is repeated.  If \code{each} is a vector, then it must has the same length as the number of columns in x}
}
\value{
A matrix with each column repeated \code{each} times
a<- matrix(1:12, 3)
repCols(a, c(2,3,4, 5))
}
\description{
repeats each column of a matrix a specified number of times
}

