% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{repRows}
\alias{repRows}
\title{repeats each row of a matrix a specified number of times}
\usage{
repRows(x, each)
}
\arguments{
\item{x}{Matrix}

\item{each}{Scalar or a vector.  This represents the number of times each row is repeated.  If \code{each} is a vector, then it must has the same length as the number of rows in x}
}
\value{
A matrix with each row repeated \code{each} times
}
\description{
repeats each row of a matrix a specified number of times
}
\examples{
a<- matrix(1:12, 3)
repRows(a, c(2,3,4))
}

