% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumRowLimited}
\alias{sumRowLimited}
\title{Calculate the limited cumulative sum of each row in a matrix}
\usage{
sumRowLimited(x, limits)
}
\arguments{
\item{x}{A numeric matrix}

\item{limits}{A vector showing the number of values to be added in each column, starting at the left side}
}
\value{
A matrix containing the cumulative product of each column of the matrix \code{x}, with limited values in each row used
}
\description{
Calculate the limited cumulative sum of each row in a matrix
}
\examples{
a <- matrix(1:9, 3)
sumRowLimited(a, limits = c(1,2,3))
}

